Banners/social-publisher/requirements.txt
# Core dependencies
aiohttp>=3.8.0
python-dotenv>=1.0.0
Pillow>=10.0.0

# Browser automation (для платформ без API)
playwright>=1.40.0

# YouTube API
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0

# Optional: For video processing
# moviepy>=1.0.3

# Optional: For enhanced async support
# asyncio>=3.4.3
